firewall CGAL behind alpha_wrap.cpp to speed up vc_add_ignore_label compiles
vc_add_ignore_label.cpp pulled the whole CGAL alpha_wrap/mesh template stack
into one 4900-line TU — ~120s to compile, ~62s of it CGAL frontend. move the
one CGAL function (classifyOuterAlphaWrap) into its own TU behind a CGAL-free
header. editing the tool now recompiles in ~39s instead of 120s; CGAL only
recompiles when alpha_wrap.cpp itself changes (rare). pure code move.
